HelloDjango/requirements.txt
asgiref==3.2.3
astroid==2.3.3
cffi==1.13.2
construct==2.5.3
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.2
django-ckeditor==5.8.0
django-crispy-forms==1.8.1
django-formtools==2.2
django-import-export==2.0
django-js-asset==1.2.2
django-mdeditor==0.1.17
django-reversion==3.0.5
et-xmlfile==1.0.1
Faker==3.0.0
future==0.18.2
get==2019.4.13
httplib2==0.9.2
isort==4.3.21
jdcal==1.4.1
jieba==0.40
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.0
openpyxl==3.0.2
pefile==2019.4.18
Pillow==7.0.0
post==2019.4.13
public==2019.4.13
pycparser==2.19
pylint==2.4.4
PyMySQL==0.9.3
python-dateutil==2.8.1
python-ptrace==0.9.4
pytz==2019.3
PyYAML==5.2
query-string==2019.4.13
request==2019.4.13
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
text-unidecode==1.3
typed-ast==1.4.0
Whoosh==2.7.4
wrapt==1.11.2
xadmin2==2.0.3
xlrd==1.2.0
xlwt==1.3.0
